+++ src/index.js
@@ -6,7 +6,7 @@ import * as audio from "./audio/audio.js";
 import collision from "./physics/collision.js";
 import * as event from "./system/event.js";
 import * as device from "./system/device.js";
-import * as game from "./game.js";
+import game from "./game.js";
 import loader from "./loader/loader.js";
 import * as Math from "./math/math.js";
 import utils from "./utils/utils.js";