+++ src/renderable/container.js
@@ -1,5 +1,5 @@
 import utils from "./../utils/utils.js";
-import * as game from "./../game.js";
+import game from "./../game.js";
 import * as event from "./../system/event.js";
 import pool from "./../system/pooling.js";
 import state from "./../state/state.js";